ref(seer): Split the overview page into focused components
index.tsx was a ~370-line god component doing feature gating, query-param
decodes, two data hooks, localStorage state, the toolbar, focus mode, and the
section list, with its content region a five-deep nested ternary interleaving
focus and list modes.

Extract three focused components alongside it: OverviewFilters (the toolbar),
FocusedIssue (the ?id= pinned-issue query and its error / pending / not-found /
card branches), and SectionList (useAutofixSections plus the per-section
rendering). index.tsx is now a thin shell that decodes the URL, owns the
shared view and collapse state, and composes the three behind the feature gate
as a single focus-vs-list branch.

Pure mechanical extraction: no behavior change. SortValue and OverviewView move
to types.ts so the pieces can share them.
